Refactored street model implementation, architecture clean up.
