+++ src/main/java/com/example/bustracker/authentication/ApplicationUser.java
@@ -1,7 +1,6 @@
 package com.example.bustracker.authentication;
 
 import org.springframework.security.core.GrantedAuthority;
-import com.example.bustracker.persistence.model.User;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import java.util.Collection;
