+++ src/main/java/com/example/bustracker/authentication/ApplicationUserService.java
@@ -1,14 +1,16 @@
 package com.example.bustracker.authentication;
 
-import com.example.bustracker.persistence.dao.ApplicationUserDAO;
+import com.example.bustracker.dao.ApplicationUserDAO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@Transactional
 public class ApplicationUserService implements UserDetailsService {
 
     @Autowired
