+++ src/main/java/com/example/bustracker/authentication/dataaccess/MySQLDataAccessService.java
@@ -1,8 +1,9 @@
-package com.example.bustracker.authentication;
+package com.example.bustracker.authentication.dataaccess;
 
-import com.example.bustracker.persistence.dao.ApplicationUserDAO;
-import com.example.bustracker.persistence.model.Role;
-import com.example.bustracker.persistence.model.User;
+import com.example.bustracker.authentication.ApplicationUser;
+import com.example.bustracker.authentication.Role;
+import com.example.bustracker.authentication.User;
+import com.example.bustracker.dao.ApplicationUserDAO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.ResultSetExtractor;
