+++ src/main/java/com/example/bustracker/city/street/StreetService.java
@@ -1,20 +1,26 @@
 package com.example.bustracker.city.street;
 
+import com.example.bustracker.city.street.dataaccess.MySQLDataAccessService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
-public class StreetService {
-    private final StreetDataAccessService streetDataAccessService;
+@Transactional
+public class StreetService implements StreetDetailsService {
+
+    private final MySQLDataAccessService mySQLDataAccessService;
 
     @Autowired
-    public StreetService(StreetDataAccessService streetDataAccessService) {
-        this.streetDataAccessService = streetDataAccessService;
+    public StreetService(@Qualifier("MySQLStreet") MySQLDataAccessService mySQLDataAccessService) {
+        this.mySQLDataAccessService = mySQLDataAccessService;
     }
 
-    List<Street> getAllStreets(){
-        return streetDataAccessService.selectStreets();
+    @Override
+    public List<Street> getAllStreets() {
+        return mySQLDataAccessService.selectStreets();
     }
 }
