+++ src/main/java/com/example/bustracker/dao/StreetDAO.java
@@ -1,0 +1,10 @@
+package com.example.bustracker.dao;
+
+
+import com.example.bustracker.city.street.Street;
+
+import java.util.List;
+
+public interface StreetDAO {
+    List<Street> selectStreets();
+}
