+++ src/main/java/com/example/bustracker/security/ApplicationMainController.java
@@ -1,4 +1,4 @@
-package com.example.bustracker.controller;
+package com.example.bustracker.security;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,13 +6,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping("/")
-public class IndexController {
+public class ApplicationMainController {
 
     @GetMapping("index")
     public String getIndexView(){
         return "index";
     }
-    
+
     @GetMapping("dashboard")
     public String getAdminView(){
         return "account/dashboard";
